Split SpeedTester.java back out into its own class
