+++ README.md
@@ -1,4 +1,4 @@
-# COMP 250: Assignment 3 - Grand Unified Tester v1.1
+# COMP 250: Assignment 3 - Grand Unified Tester v1.2
 This repository is meant to act as a bundle of
 several tools and testers others have 
 shared over the past days and weeks. 
@@ -11,16 +11,16 @@ Includes:
 - [ZTestCode.java](https://github.com/Lprob1/Assignment3_COMP250)
     * Lprob1's test cases
 - TreePrinter.java
-    * Prints a binary tree - you just need to temporarily make DogNode implement
-      PrintableNode - just make sure you reverse it after!
+    * Prints a binary tree - you just need to temporarily make DogNode implement PrintableNode - just make sure you
+      reverse it after!
     * from [michal.kreuzman](https://stackoverflow.com/a/4973083) on StackOverflow
- - __M E G A T E S T E R !__ Includes:
-   * SpeedTester.java
-     * speedtesting functions by huubdejong 
-     * Requires TheBigSasha's [RuntimeTester](https://github.com/TheBigSasha/RuntimeTester)
-   * [DogShelterTest.java](https://github.com/jamesxu123/COMP-250-A3-Tests)
-     * jamesxu123's test cases as well as a couple other tester methods
-   * Minitester tests, modified to use JUnit Assertions so they can be run along with all the other tests in IntelliJ
+- SpeedTester.java
+    * speedtesting functions by huubdejong
+    * Requires TheBigSasha's [RuntimeTester](https://github.com/TheBigSasha/RuntimeTester)
+- __M E G A T E S T E R !__ Includes:
+    * [DogShelterTest.java](https://github.com/jamesxu123/COMP-250-A3-Tests)
+        * jamesxu123's test cases as well as a couple other tester methods
+    * Minitester tests, modified to use JUnit Assertions so they can be run along with all the other tests in IntelliJ
     
 and more!
 
@@ -36,22 +36,29 @@ Eclipse
   4. Select the "master" branch
 
 ## Usage
-All instructions are written with IntelliJ in mind; also not sure about JUnit's compatibility with Eclipse out-of-the-box, so yeah try IntelliJ.
+
+Now supports Eclipse! You should be able to import the file and just All instructions are written with IntelliJ in mind;
+also not sure about JUnit's compatibility with Eclipse out-of-the-box, so yeah try IntelliJ.
+
 * First of all, add your DogShelter code in place of the blank/template DogShelter file.
-* DViz: 
+* DViz:
     * Run `DViz.java`.
 * __M E G A T E S T E R__:
     * Test cases:
-       * (In IntelliJ) right click the folder `assignment3` under `src` and click "Run Tests in 'assignment3", or select "Run..." on Megatester.java through the menu or though Alt-Shift-F10 (or the Mac equivalent), and select the one with orange and green arrows
-       * Alternatively, comment out `Visualizer.launch(Megatester.java)` in `main()`, uncomment the rest of `main()`, and run Megatester.java like you normally would
-     * SpeedTester: 
-       * Follow the instructions on [this page](https://github.com/TheBigSasha/RuntimeTester) to download RuntimeTester.jar and add it to your project.
-       * Run Megatester.java like you normally would
+        * (In IntelliJ) right click the folder `assignment3` under `src` and click "Run Tests in 'assignment3", or
+          select "Run..." on Megatester.java through the menu or though Alt-Shift-F10 (or the Mac equivalent Ctrl-R),
+          and select the one with orange and green arrows
+
+* SpeedTester:
+    * Follow the instructions on [this page](https://github.com/TheBigSasha/RuntimeTester) to download RuntimeTester.jar
+      and add it to your project.
+    * Run SpeedTester.java.
 * TreePrinter
     * Make DogNode implement PrintableNode, including its 3 methods. (remove this after!)
     * Call `TreePrinter.print(shelter.root)` on the root of the desired shelter.
 * Lprob1's tests
-    * Run ZTestCode.java like you would normally. These tests don't user JUnit - Lprob1 intended them to be such that you determine whether or not the results are correct based on what's printed to the console.
+    * Run ZTestCode.java like you would normally. These tests don't user JUnit - Lprob1 intended them to be such that
+      you determine whether or not the results are correct based on what's printed to the console.
 
 ## Contributing
 Feel free to fork and make a pull request if you have any tests you think could be added!
